doc (utime_ex): 添加utime_ex的README和API手册
添加utime_ex的README和API手册，英文未做校对

固件版本: N/A

是否需要文案翻译: 是
