+++ README_ZH.md
@@ -1,0 +1,33 @@
+# QuecPython 增强型时间及时区
+
+中文 | [English](README.md)
+
+## 概述
+
+QuecPython utime方法只能支持东西十二时区，但在实际应用中，某些国境横跨国际日期变更线的国家为保持全国日期统一,将原属UTC-11的领土日期前推一天,即从UTC-11加24H,是为UTC+13。该方法可以拓展时区至为UTC-12至UTC+14。
+
+同时，utime方法没有便捷的设定时间的方法，需要另外调用RTC中设置时间的API，并且需要对日期格式做转换，本方法提供格式一致的获取/设置本地时间的方法。
+
+最后，本方法提供了更为灵活的mktime方法，可以增加加时区参数，直接获取指定时区的时间戳。
+
+## 用法
+
+- [API 参考手册](./docs/zh/API参考手册.md)
+
+## 贡献
+
+我们欢迎对本项目的改进做出贡献！请按照以下步骤进行贡献：
+
+1. Fork 此仓库。
+2. 创建一个新分支（`git checkout -b feature/your-feature`）。
+3. 提交您的更改（`git commit -m 'Add your feature'`）。
+4. 推送到分支（`git push origin feature/your-feature`）。
+5. 打开一个 Pull Request。
+
+## 许可证
+
+本项目使用 Apache 许可证。详细信息请参阅 [LICENSE](LICENSE) 文件。
+
+## 支持
+
+如果您有任何问题或需要支持，请参阅 [QuecPython 文档](https://python.quectel.com/doc) 或在本仓库中打开一个 issue。
